move protocolAddresses.json to a generated folder
